CASSANDRA-16066: Add website generation tooling
This commit replaces the existing Docker website generation tool. The original
website generation tool was Jekyll. It has been replaced with Antora which
is purpose built for handling different document versions in a repository.

A Docker container has been added which renders the site using Antora. Unlike
other Antora configurations which have a site.yaml file committed, this
implementaiton uses a site.yaml template and python script to generate the
final site.yaml. The Docker container is responsible for calling the script
to generate the site.yaml file using the template. The generated file is then
passed to Antora to render the final site. This has been done so that different
document sources can be passed to Antora and newly released Cassandra versions
will automatically appear in the Downloads page.

The source content and styling which Antora uses to generate the site can be
controlled via the container environment variables.

The container includes a preview mode which monitors the content directories
and runs Antora when any file changes in the content directories.

By default, only the cassandra-website will be used as the source for Antora.
That is unless the Cassandra document generation is specified via the
generate-docs command, the container will only generate the website HTML.

patch by Anthony Grasso; reviewed by Mick Semb Wever, Lorina Poland, Melissa Logan, Paul Au for CASSANDRA-16066