CASSANDRA-16066: Set up new directory structure
This commit only sets up the new directory structure. Content, styling/layout,
and tooling changes will appear in seperate commits.

The old directory structure was:

ROOT
  - content
  - src
    - _data
    - _includes
    - _layouts
    - _plugins
    - _posts
    - _sass
    - _templates
    - blog
    - css
    - doc
    - icons
    - img
    - js

In the above old structure the styling and the content all lived in the same
parent directory. The new directory structure separates the styling from the
content. The new directory structure is now:

 ROOT
   - site-content
   - site-ui

The 'site-ui' directory will contain only the styling that determines the look
and feel of the site. A ui-bundle.zip file containing the styling information
will be generated using the contents of the 'site-ui' directory. Generation of
the ui-bundle.zip will be done using Gulp running inside a Docker container.

The 'site-content' directory will contain all the raw page information e.g.
where to download, developer guidelines, how to commit patches, etc. The final
website HTML that is deployed to production is generated from this directory.
Generation of the HTML content will be performed by Antora running inside a
Docker container. As part of the website HTML generation, the ui-bundle.zip
file and the Cassandra documentation location are passed to Antora. Antora uses
the ui-bundle.zip to style the website. The Cassandra documentation location
will be used to gather and generate documentation for each Cassandra version.

Antora is being used for the website generation because it is designed to create
websites that have version documentation. For example, when a new version of
Cassandra is released, a new version of the documentation will be generated as
well. Hence, if there is a change in the behaviour of Cassandra or a tool in the
project, it will be captured in the latest version of the documentation. Users
of the project will have the ability to select the version of the documentation
they are interested in.

Separating the layout/style and the content means that stying changes can be
made with little to no impact on content and vice-versa. In addition, changes to
the styling can happen in parallel while website content is updated without
conflict.

The markdown files contained in the '_post' old structure will be converted to
asciidoc and placed 'site-content' directory. In addition, various markdown
pages that live in the old ROOT directory will be converted to asciidoc and
placed 'site-content' directory as well.