CASSANDRA-16066: Add Example site-ui components
Added to the 'site-ui' directory an example an Antora site UI implementation.
The example was taken from https://gitlab.com/antora/antora-ui-default. It
shows the styling components that are needed in the 'site-ui' and how the
ui-bundle.zip can be generated.

The associated tooling to call Gulp to preview, build and generate the
ui-bundle.zip will be added in another commit.