CASSANDRA-16066: Add new website content and styling
This commit adds the content files and styling resources to render the new
website.

The content added are ADOC files and associated images to the following
directory. It is placed in this directory to be consumed by Antora which is used to
generate the website HTML.

    site-content/source/modules/ROOT/pages/

The styling resources added are CSS, JS, layouts, and partials to the following
directory. It is placed in this directory to be consumed by gulp which will
package the resources into a ZIP file. The ZIP file is then passed to Antora
when the site is rendered.

    site-ui/src/

patch by Paul Au; reviewed by Anthony Grasso, Erick Ramirez, Mick Semb Wever for CASSANDRA-16762