CASSANDRA-17274: Fix link syntax community component
Previously, link for contribute to cassandra was broken and rendered as
plaintext instead of a link tag, this was because the link was a
relative instead of absolute link and URL macro was not able to
understand that.

Using link macro solved the issue and corrected the path to
development/index.html for contribute to cassandra.